refactor: rename `functions_loader`
